Improve user specific configuration
Do not consider /home as the only folder where home directories
might be located. To find all possible home locations, we try to
login for each user listed in variable powerline_users. Finally
echo $HOME is executed in the context of the user session. This
aproach should work for all kind of possible user setups.

Further on, this works as well for domain users who have never been
logged in to a hosts. During login the home directories for all
users listed in variable powerline are created.